Fixing clean script to include git capabilities.

+++ clean.sh
@@ -1,12 +1,21 @@
 #!/bin/bash
+
 echo
 echo "The following files will be purged:"
-echo
-svn status --no-ignore | grep ^[I?] | sed 's/^[I?]//g' | sed 's/^[ ].//g'
+if [ -d .svn ]; then
+	svn status --no-ignore | grep ^[I?] | sed 's/^[I?]//g' | sed 's/^[ ].//g'
+else
+	git-clean -n -f -x -d
+fi
+
 echo
 read -p "Continue (Y/N)?"
 if [ "$REPLY" == "y" ]; then
-	rm -rvf `svn status --no-ignore | grep ^[I?] | sed 's/^[I?]//g' | sed 's/^[ ].//g'`
+	if [ -d .svn ]; then
+		rm -rvf `svn status --no-ignore | grep ^[I?] | sed 's/^[I?]//g' | sed 's/^[ ].//g'`
+	else
+		git-clean -f -x -d
+	fi
 else
 	echo
 	echo "No files removed."
